Clean headers & reorganize backend agnostic code

+++ src/backend/opengl/font_impl.cpp
@@ -7,18 +7,16 @@
  * http://arrayfire.com/licenses/BSD-3-Clause
  ********************************************************/
 
-#include <common.hpp>
+#include <algorithm>
+#include <cmath>
+#include <common/err_handling.hpp>
+#include <cstring>
 #include <font_impl.hpp>
+#include <gl_helpers.hpp>
+#include <glm/gtc/matrix_transform.hpp>
 #include <shader_headers/font_vs.hpp>
 #include <shader_headers/font_fs.hpp>
-
-#include <cmath>
-#include <cstring>
 #include <sstream>
-#include <algorithm>
-
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
 
 #include <ft2build.h>
 #include FT_FREETYPE_H
@@ -54,10 +52,8 @@ static const struct {
 
 /* freetype library types */
 
-namespace forge
-{
-namespace opengl
-{
+namespace forge {
+namespace opengl {
 
 const float MIN_FONT_SIZE = 8.0f;
 const float MAX_FONT_SIZE = 24.0f;
@@ -265,7 +261,7 @@ font_impl::~font_impl()
     mAtlas.reset();
 }
 
-void font_impl::setOthro2D(int pWidth, int pHeight)
+void font_impl::setOthro2D(size_t pWidth, size_t pHeight)
 {
     mOrthoW  = pWidth;
     mOrthoH  = pHeight;
@@ -300,7 +296,7 @@ void font_impl::loadFont(const char* const pFile)
     /* push each glyphs vertex and texture data into VBO */
     std::vector<float> vdata;
 
-    uint index = 0;
+    uint32_t index = 0;
 
     for (size_t f=0; f<mGlyphLists.size(); ++f)
     {
